requirements.txt
python-docx==1.1.0
flask==2.3.2